requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
